backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-jwt-extended>=4.6.0
flask-limiter>=3.5.0
prometheus-flask-exporter>=0.23.0
sqlalchemy>=2.0.0
alembic>=1.13.0
werkzeug>=3.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
xgboost==2.0.3
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
joblib==1.4.2
reportlab==4.2.5
Pillow==11.0.0
pytest==8.3.3
requests==2.32.3
